bug fixed
