+++ index.html
@@ -63,6 +63,6 @@
       </section>
     </main>
     <script src="./index.js" type="module"></script>
-    <script src="./modules/reg_log_auth.js" type="module"></script>
+    <script src="./modules/signInAuth.js" type="module"></script>
   </body>
 </html>
