+++ register.html
@@ -71,7 +71,7 @@
             <i class="bx bx-lock-alt"></i>
           </div>
           <div class="input-box">
-            <button class="submit" type="submit" id="registerButton">
+            <button id="registerButton" class="submit" type="submit">
               Register
             </button>
           </div>
@@ -88,6 +88,6 @@
       </form>
     </main>
     <script src="./index.js" type="module"></script>
-    <script src="./modules/reg_log_auth.js" type="module"></script>
+    <script src="./modules/registerAuth.js" type="module"></script>
   </body>
 </html>
